requirements.txt
pdfplumber
langchain
langchain-huggingface
langchain-community
pypdf
langchain_chroma
chromadb
smolagents
python-dotenv
gradio
sentence-transformers
openai
tiktoken